Changed identification to use HSV
RGB analysis was not brightness resistant enough
HSV seems to do a better job. The problem still is the huge amount of trees that I try to filter out by blacking all pixels that are green enough (in the green spectrum of the H channel).
The resulting image and its negative seem to do a good job of retaining the signs where I can extract it with contour finding. OpenCVs implementation doesn't seem to like small round objects so that I still have to look into